add the correct files
