+++ readme.md
@@ -1,0 +1,63 @@
+![](https://img.shields.io/badge/Microverse-blueviolet)
+
+# To-Do List;
+
+> In this project, we built a basic website that allows users to add/remove books from a list.We achieved that by using JavaScript objects and arrays.we also dynamically modified the DOM and add basic events.
+
+
+## Built With
+
+- Javascript ES6
+- HTML and CSS
+- webpack
+  
+
+## Live Demo (if available)
+
+[Incoming ...](#)
+
+## Getting Started
+
+
+To get a local copy up and running follow these simple example steps.
+
+clone the project by using terminal: 
+
+`git@github.com:MohamedCK/To-Do-List.git`
+
+### prerequisites
+- Javascript
+- webpack
+### setup
+- clone or download this repo on your machine.
+- Enter the project directory,
+- Run `npm install`,
+- Then run  `npm start` to check the live server
+
+## Authors
+
+👤 **Mohamed Abdulhakim**
+
+- GitHub: [@MohamedCK](https://github.com/MohamedCK)
+- Twitter: [@MohamedCK](https://twitter.com/MohamedCK0)
+- LinkedIn: [@MohamedCK](https://www.linkedin.com/in/mohamed-abdulhakim-2868521b6/)
+
+## 🤝 Contributing
+
+Contributions, issues, and feature requests are welcome!
+
+Feel free to check the [issues page](https://github.com/MohamedCK/Awesome-books-plain-JavaScript-with-objects/issues).
+
+## Show your support
+
+Give a ⭐️ if you like this project!
+
+## Acknowledgments
+
+- Hat tip to anyone whose code was used
+- Inspiration
+- etc
+
+## 📝 License
+
+This project is [MIT](./MIT.md) licensed.
