+++ src/template.html
@@ -6,7 +6,6 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>To Do List</title>
-  <link rel="stylesheet" href="style.css">
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.2/css/all.min.css">
 </head>
 
@@ -32,7 +31,7 @@
       <button class="clearall" type="button">Clear all completed</button>
     </div>
   </div>
-  <script type = "module" src="index.js"></script>
 </body>
 
 </html>
+
