--- style.css
@@ -1,75 +1,0 @@
-body {
-    font-family: sans-serif;
-    background-color: #f45b69;
-    width: 60%;
-    margin: 100px auto;
-    box-shadow: -10px 10px 8px 1px rgba(0, 0, 0, 0.19), 15px 10px 20px 0 rgba(0, 0, 0, 0.19);
-  }
-  
-  .container {
-    background-color: #ffff;
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .title {
-    padding: 0 10px;
-    display: grid;
-    grid-template-columns: 10fr 1fr;
-    border-bottom: 2px solid #f0f0f0;
-  }
-  
-  .refresh {
-    background: none;
-    border: none;
-  }
-  
-  .input-container {
-    padding: 0 10px;
-    display: grid;
-    grid-template-columns: 10fr 1fr;
-    border-bottom: 2px solid #f0f0f0;
-  }
-  
-  .input {
-    border: none;
-    line-height: 50px;
-  }
-  
-  .submit {
-    border: none;
-    background: none;
-  }
-  
-  .check {
-    padding: 0 10px;
-    display: grid;
-    grid-template-columns: 0.5fr 8fr 1fr;
-    align-items: center;
-    text-align: center;
-    height: 50px;
-    border-bottom: 1px solid #f6f6f6;
-  }
-  
-  .input-check {
-    height: 30px;
-    border: 2px solid #bdbdbd;
-  }
-  
-  .edit-list {
-    border: none;
-    padding: 12px;
-  }
-  
-  .clear-container {
-    width: 100%;
-  }
-  
-  .clearall {
-    background: #f6f6f6;
-    border: none;
-    width: 100%;
-    height: 50px;
-    text-align: center;
-    color: #ababab;
-  }
